requirements.txt
streamlit>=1.28.0
pydantic>=2.0.0

llama-index>=0.9.0
llama-index-llms-huggingface>=0.1.0
llama-index-embeddings-huggingface>=0.1.0
llama-index-vector-stores-chroma>=0.1.0

chromadb>=0.4.0

datasets>=2.14.0
transformers>=4.35.0
accelerate>=0.24.0
bitsandbytes>=0.41.0
sentence-transformers>=2.2.0

torch>=2.0.0

requests>=2.31.0
pyyaml>=6.0
tenacity>=8.2.0

python-json-logger>=2.0.0

protobuf>=4.24.0